Rename file for RetryProcessor class
