+++ src/functions/retryProcessing/application/IRetryProcessor.ts
@@ -1,16 +1,22 @@
 
 export interface IRetryProcessor {
   processRetries(): Promise<void>;
+
   processSuccessful(): Promise<void>;
+
   processErrorsToRetry(
     rsisRetryCount: number,
     notifyRetryCount: number,
-    tarsRetryCount: number): Promise<void>;
+    tarsRetryCount: number,
+  ): Promise<void>;
+
   processErrorsToAbort(
     rsisRetryCount: number,
     notifyRetryCount: number,
-    tarsRetryCount: number): Promise<void>;
-  processSupportInterventions(): Promise<void>;
-  processOldEntryCleanup(cutOffPointInDays: number): Promise<void>;
+    tarsRetryCount: number,
+  ): Promise<void>;
 
+  processSupportInterventions(): Promise<void>;
+
+  processOldEntryCleanup(cutOffPointInDays: number): Promise<void>;
 }
