+++ src/functions/retryProcessing/framework/handler.ts
@@ -2,7 +2,7 @@ import { APIGatewayEvent, Context } from 'aws-lambda';
 import createResponse from '../../../common/application/utils/createResponse';
 import { HttpStatus } from '../../../common/application/api/HttpStatus';
 import Response from '../../../common/application/api/Response';
-import { RetryProcessor } from '../application/processRetries';
+import { RetryProcessor } from '../application/RetryProcessor';
 import { bootstrapConfig } from '../../../../src/common/framework/config/config';
 import { IRetryProcessor } from '../application/IRetryProcessor';
 import { error } from '@dvsa/mes-microservice-common/application/utils/logger';
